Add files via upload
The python has been updated to work with Python2.7. I also fixed an issue where we experienced wrong password because a newline character was being added to the password during login. I can confirm this works with Teradata 14.x.x and Teradata 15.x.x at this time. There are some missing character types used in the newer version of Teradata this does not have yet. Will look at adding them in the future.